Upgrade component
